rename check-form-v2 controller to test-developer2
