+++ admin/spec/back-end/controllers/test-developer2.spec.js
@@ -3,12 +3,12 @@
 /* global describe beforeEach it expect jasmine spyOn */
 
 const httpMocks = require('node-mocks-http')
-const controller = require('../../../controllers/check-form-v2')
+const sut = require('../../../controllers/test-developer2')
 const checkFormPresenter = require('../../../helpers/check-form-presenter')
 const checkFormV2Service = require('../../../services/check-form-v2.service')
 const checkWindowV2Service = require('../../../services/check-window-v2.service')
 
-describe('check form v2 controller:', () => {
+describe('test developer 2 controller:', () => {
   let next
   beforeEach(() => {
     next = jasmine.createSpy('next')
@@ -38,7 +38,7 @@ describe('check form v2 controller:', () => {
       const req = getReq(reqParams)
       spyOn(res, 'render')
       spyOn(checkFormV2Service, 'getSavedForms')
-      await controller.getViewFormsPage(req, res, next)
+      await sut.getViewFormsPage(req, res, next)
       expect(res.locals.pageTitle).toBe('Upload and view forms')
       expect(checkFormV2Service.getSavedForms).toHaveBeenCalled()
       expect(res.render).toHaveBeenCalled()
@@ -54,7 +54,7 @@ describe('check form v2 controller:', () => {
       const req = getReq(reqParams)
       spyOn(checkFormV2Service, 'hasExistingFamiliarisationCheckForm')
       spyOn(res, 'render')
-      await controller.getUploadNewFormsPage(req, res, next)
+      await sut.getUploadNewFormsPage(req, res, next)
       expect(res.locals.pageTitle).toBe('Upload new form')
       expect(res.render).toHaveBeenCalled()
     })
@@ -64,7 +64,7 @@ describe('check form v2 controller:', () => {
       const error = new Error('error')
       spyOn(checkFormV2Service, 'hasExistingFamiliarisationCheckForm').and.returnValue(Promise.reject(error))
       spyOn(res, 'render')
-      await controller.getUploadNewFormsPage(req, res, next)
+      await sut.getUploadNewFormsPage(req, res, next)
       expect(res.render).not.toHaveBeenCalled()
       expect(next).toHaveBeenCalledWith(error)
     })
@@ -86,7 +86,7 @@ describe('check form v2 controller:', () => {
       spyOn(res, 'redirect')
       spyOn(checkFormV2Service, 'saveCheckForms')
       spyOn(checkFormPresenter, 'getFlashMessageData')
-      await controller.postUpload(req, res, next)
+      await sut.postUpload(req, res, next)
       expect(checkFormV2Service.saveCheckForms).toHaveBeenCalled()
       expect(res.redirect).toHaveBeenCalled()
     })
@@ -97,7 +97,7 @@ describe('check form v2 controller:', () => {
       const error = new Error('error')
       spyOn(checkFormV2Service, 'saveCheckForms').and.returnValue(Promise.reject(error))
       spyOn(checkFormPresenter, 'getFlashMessageData')
-      await controller.postUpload(req, res, next)
+      await sut.postUpload(req, res, next)
       expect(checkFormV2Service.saveCheckForms).toHaveBeenCalled()
       expect(checkFormPresenter.getFlashMessageData).not.toHaveBeenCalled()
       expect(res.redirect).not.toHaveBeenCalled()
@@ -118,7 +118,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormV2Service, 'getCheckFormName')
       spyOn(checkFormV2Service, 'deleteCheckForm')
       spyOn(res, 'redirect')
-      await controller.getDelete(req, res, next)
+      await sut.getDelete(req, res, next)
       expect(checkFormV2Service.getCheckFormName).toHaveBeenCalled()
       expect(checkFormV2Service.deleteCheckForm).toHaveBeenCalled()
       expect(res.redirect).toHaveBeenCalled()
@@ -130,7 +130,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormV2Service, 'getCheckFormName')
       spyOn(checkFormV2Service, 'deleteCheckForm').and.returnValue(Promise.reject(error))
       spyOn(res, 'redirect')
-      await controller.getDelete(req, res, next)
+      await sut.getDelete(req, res, next)
       expect(checkFormV2Service.getCheckFormName).toHaveBeenCalled()
       expect(checkFormV2Service.deleteCheckForm).toHaveBeenCalled()
       expect(res.redirect).not.toHaveBeenCalled()
@@ -150,7 +150,7 @@ describe('check form v2 controller:', () => {
       const req = getReq(reqParams)
       spyOn(checkFormV2Service, 'getCheckForm').and.returnValue({ checkFormName: 'checkFormName' })
       spyOn(res, 'render')
-      await controller.getViewFormPage(req, res, next)
+      await sut.getViewFormPage(req, res, next)
       expect(checkFormV2Service.getCheckForm).toHaveBeenCalled()
       expect(res.locals.pageTitle).toBe('checkFormName')
       expect(res.render).toHaveBeenCalled()
@@ -161,7 +161,7 @@ describe('check form v2 controller:', () => {
       const error = new Error('error')
       spyOn(checkFormV2Service, 'getCheckForm').and.returnValue(Promise.reject(error))
       spyOn(res, 'render')
-      await controller.getViewFormPage(req, res, next)
+      await sut.getViewFormPage(req, res, next)
       expect(checkFormV2Service.getCheckForm).toHaveBeenCalled()
       expect(res.render).not.toHaveBeenCalled()
       expect(next).toHaveBeenCalledWith(error)
@@ -178,7 +178,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkWindowV2Service, 'getPresentAndFutureCheckWindows')
       spyOn(checkFormPresenter, 'getPresentationCheckWindowListData')
       spyOn(res, 'render')
-      await controller.getAssignFormsPage(req, res, next)
+      await sut.getAssignFormsPage(req, res, next)
       expect(checkWindowV2Service.getPresentAndFutureCheckWindows).toHaveBeenCalled()
       expect(checkFormPresenter.getPresentationCheckWindowListData).toHaveBeenCalled()
       expect(res.locals.pageTitle).toBe('Assign forms to check window')
@@ -191,7 +191,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkWindowV2Service, 'getPresentAndFutureCheckWindows').and.returnValue(Promise.reject(error))
       spyOn(checkFormPresenter, 'getPresentationCheckWindowListData')
       spyOn(res, 'render')
-      await controller.getAssignFormsPage(req, res, next)
+      await sut.getAssignFormsPage(req, res, next)
       expect(checkWindowV2Service.getPresentAndFutureCheckWindows).toHaveBeenCalled()
       expect(checkFormPresenter.getPresentationCheckWindowListData).not.toHaveBeenCalled()
       expect(res.render).not.toHaveBeenCalled()
@@ -216,7 +216,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormPresenter, 'getPresentationCheckWindowData').and.returnValue({ name: 'checkWindowName', checkPeriod: 'MTC' })
       spyOn(checkFormPresenter, 'getPresentationAvailableFormsData')
       spyOn(res, 'render')
-      await controller.getSelectFormPage(req, res, next)
+      await sut.getSelectFormPage(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByType).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByCheckWindowIdAndType).toHaveBeenCalled()
@@ -235,7 +235,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormPresenter, 'getPresentationCheckWindowData')
       spyOn(checkFormPresenter, 'getPresentationAvailableFormsData')
       spyOn(res, 'render')
-      await controller.getSelectFormPage(req, res, next)
+      await sut.getSelectFormPage(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByType).not.toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByCheckWindowIdAndType).not.toHaveBeenCalled()
@@ -254,7 +254,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormPresenter, 'getPresentationCheckWindowData')
       spyOn(checkFormPresenter, 'getPresentationAvailableFormsData')
       spyOn(res, 'render')
-      await controller.getSelectFormPage(req, res, next)
+      await sut.getSelectFormPage(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByType).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByCheckWindowIdAndType).not.toHaveBeenCalled()
@@ -273,7 +273,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormPresenter, 'getPresentationCheckWindowData')
       spyOn(checkFormPresenter, 'getPresentationAvailableFormsData')
       spyOn(res, 'render')
-      await controller.getSelectFormPage(req, res, next)
+      await sut.getSelectFormPage(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByType).toHaveBeenCalled()
       expect(checkFormV2Service.getCheckFormsByCheckWindowIdAndType).toHaveBeenCalled()
@@ -320,7 +320,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormV2Service, 'hasAssignedFamiliarisationForm').and.returnValue(true)
       spyOn(checkFormV2Service, 'updateCheckWindowForms')
       spyOn(checkFormPresenter, 'getAssignFormsFlashMessage')
-      await controller.postAssignForms(req, res, next)
+      await sut.postAssignForms(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.hasAssignedFamiliarisationForm).toHaveBeenCalled()
       expect(checkFormV2Service.updateCheckWindowForms).toHaveBeenCalled()
@@ -336,7 +336,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormV2Service, 'hasAssignedFamiliarisationForm')
       spyOn(checkFormV2Service, 'updateCheckWindowForms')
       spyOn(checkFormPresenter, 'getAssignFormsFlashMessage')
-      await controller.postAssignForms(req, res, next)
+      await sut.postAssignForms(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.hasAssignedFamiliarisationForm).not.toHaveBeenCalled()
       expect(checkFormV2Service.updateCheckWindowForms).not.toHaveBeenCalled()
@@ -353,7 +353,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormV2Service, 'hasAssignedFamiliarisationForm').and.returnValue(true)
       spyOn(checkFormV2Service, 'updateCheckWindowForms').and.returnValue(Promise.reject(error))
       spyOn(checkFormPresenter, 'getAssignFormsFlashMessage')
-      await controller.postAssignForms(req, res, next)
+      await sut.postAssignForms(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.hasAssignedFamiliarisationForm).toHaveBeenCalled()
       expect(checkFormV2Service.updateCheckWindowForms).toHaveBeenCalled()
@@ -369,7 +369,7 @@ describe('check form v2 controller:', () => {
       spyOn(checkFormV2Service, 'hasAssignedFamiliarisationForm').and.returnValue(false)
       spyOn(checkFormV2Service, 'updateCheckWindowForms')
       spyOn(checkFormPresenter, 'getAssignFormsFlashMessage')
-      await controller.postAssignForms(req, res, next)
+      await sut.postAssignForms(req, res, next)
       expect(checkWindowV2Service.getCheckWindow).toHaveBeenCalled()
       expect(checkFormV2Service.hasAssignedFamiliarisationForm).toHaveBeenCalled()
       expect(checkFormV2Service.updateCheckWindowForms).not.toHaveBeenCalled()
